generated_projects/test_integration/requirements.txt
fastapi
sqlalchemy
bcrypt
pyjwt
uvicorn